% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_blank_address_layer.R
\name{add_blank_address_layer}
\alias{add_blank_address_layer}
\title{Add a blank address layer}
\usage{
add_blank_address_layer(map)
}
\arguments{
\item{map}{Map created via \link{map_toronto}}
}
\description{
Add an empty layer of \link{address_points} to a map (created via \link{map_toronto}). The purpose of this function is to allow for toggling highlighting an address, via \link{zoom_map_to_address}.
}
\examples{
library(sf)

map_toronto() \%>\%
  add_blank_address_layer()
}
