% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_blank_layers.R
\name{add_blank_amenity_density_layer}
\alias{add_blank_amenity_density_layer}
\title{Add a blank amenity density layer}
\usage{
add_blank_amenity_density_layer(map)
}
\arguments{
\item{map}{Map created via \link{map_toronto}}
}
\description{
Add an empty layer of \link{amenity_density} to a map (created via \link{map_toronto}). The purpose of this function is to allow for toggling the layer on and off, via \link{toggle_layer_visible} and \link{toggle_layer_invisible} (with the id "amenity_density").
}
\examples{
library(sf)

map_toronto() \%>\%
  add_blank_amenity_density_layer() \%>\%
  toggle_layer_visible("amenity_density")
}
