% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_blank_apartment_evaluation_layer.R
\name{add_blank_apartment_evaluation_layer}
\alias{add_blank_apartment_evaluation_layer}
\title{Add a blank apartment evaluation layer}
\usage{
add_blank_apartment_evaluation_layer(map)
}
\arguments{
\item{map}{Map created via \link{map_toronto}}
}
\description{
Add an invisible layer of \link{apartment_evaluation} to a map (created via \link{map_toronto}). The purpose of this function is to allow for toggling the apartment building layer on and off, via \link{toggle_layer_visible} and \link{toggle_layer_invisible}.
}
\examples{
library(sf)

map_toronto() \%>\%
  add_blank_apartment_evaluation_layer() \%>\%
  toggle_layer_visible("apartment_evaluation")
}
