% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_blank_apartment_layer.R
\name{add_blank_apartment_layer}
\alias{add_blank_apartment_layer}
\title{Add a blank apartment layer}
\usage{
add_blank_apartment_layer(map)
}
\arguments{
\item{map}{Map created via \link{map_toronto}}
}
\description{
Add a blank layer of \link{apartment_building_registry} to a map (created via \link{map_toronto}). The purpose of this function is to allow for zooming and highlighting to an address from the data, via \link{zoom_map_to_address}.
}
\examples{
library(sf)

map_toronto() \%>\%
  add_blank_apartment_layer() \%>\%
  zoom_map_to_address("378 Markham St")
}
