% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_blank_layers.R
\name{add_blank_core_housing_need_layer}
\alias{add_blank_core_housing_need_layer}
\title{Add blank core housing need layer}
\usage{
add_blank_core_housing_need_layer(map)
}
\arguments{
\item{map}{Map created via \link{map_toronto}}
}
\description{
Add a blank layers of \link{core_housing_need_by_neighbourhood} to a map (created via \link{map_toronto}).The purpose of this function is to allow for toggling the layers on and off, via \link{toggle_layer_visible} and \link{toggle_layer_invisible} (with id "core_housing_need").
}
\examples{
library(sf)
map_toronto() \%>\%
  add_blank_core_housing_need_layer() \%>\%
  toggle_layer_visible("core_housing_need")
}
