% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_blank_evictions_hearings_layer.R
\name{add_blank_evictions_hearings_layer}
\alias{add_blank_evictions_hearings_layer}
\title{Add a blank evictions hearings layer}
\usage{
add_blank_evictions_hearings_layer(map)
}
\arguments{
\item{map}{Map created via \link{map_toronto}}
}
\description{
Add an empty layer of \link{apartment_buildings} (with evictions hearings only) to a map (created via \link{map_toronto}). The purpose of this function is to allow for toggling the layer on and off, via \link{toggle_layer_visible} and \link{toggle_layer_invisible}.
}
\examples{
library(sf)

map_toronto() \%>\%
  add_blank_evictions_hearings_layer() \%>\%
  toggle_layer_visible("evictions_hearings")
}
