% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_blank_layers.R
\name{add_blank_evictions_layer}
\alias{add_blank_evictions_layer}
\title{Add blank evictions layer}
\usage{
add_blank_evictions_layer(map)
}
\arguments{
\item{map}{Map created via \link{map_toronto}}
}
\description{
Add a blank layers of \link{evictions_by_neighbourhood} to a map (created via \link{map_toronto}).The purpose of this function is to allow for toggling the layers on and off, via \link{toggle_layer_visible} and \link{toggle_layer_invisible} (with id "evictions").
}
\examples{
library(sf)
map_toronto() \%>\%
  add_blank_evictions_layer() \%>\%
  toggle_layer_visible("evictions")
}
