% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_blank_lem_layer.R
\name{add_blank_lem_layer}
\alias{add_blank_lem_layer}
\title{Add a blank LEM layer}
\usage{
add_blank_lem_layer(map)
}
\arguments{
\item{map}{Map created via \link{map_toronto}}
}
\description{
Add an empty layer of \link{total_affordable_by_neighbourhood} to a map (created via \link{map_toronto}). The purpose of this function is to allow for toggling the layer on and off, via \link{toggle_layer_visible} and \link{toggle_layer_invisible}.
}
\examples{
library(sf)

map_toronto() \%>\%
  add_blank_lem_layer() \%>\%
  toggle_layer_visible("lem")
}
