% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_blank_layers.R
\name{add_blank_neighbourhood_layer}
\alias{add_blank_neighbourhood_layer}
\title{Add a blank neighbourhood layer}
\usage{
add_blank_neighbourhood_layer(map)
}
\arguments{
\item{map}{Map created via \link{map_toronto}}
}
\description{
Add a blank layer of \link{neighbourhoods} to a map (created via \link{map_toronto}). The purpose of this function is to allow for zooming and highlighting to a neighbourhood from the data, via \link{zoom_map_to_neighbourhood}.
}
\examples{
library(sf)
map_toronto() \%>\%
  add_blank_neighbourhood_layer() \%>\%
  zoom_map_to_neighbourhood("Casa Loma")
}
