% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_blank_layers.R
\name{add_blank_points_layers}
\alias{add_blank_points_layers}
\title{Add blank points layers}
\usage{
add_blank_points_layers(map)
}
\arguments{
\item{map}{Map created via \link{map_toronto}}
}
\description{
Add empty layers of \link{buildings}. The purpose of this function is to allow for toggling the layers on and off, via \link{toggle_layer_visible} and \link{toggle_layer_invisible}. This function adds the following layers (accessed via IDs): apartment building registry (apartment_buildings), RentSafeTO scores (apartment_evaluation), evictions hearings (evictions_hearings), Above Guideline Increase applications (agi), and tenant defense fund grants (tdf).
}
\examples{
map_toronto() \%>\%
  add_blank_points_layers() \%>\%
  toggle_layer_visible("tdf")
}
