% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_blank_layers.R
\name{add_blank_rental_supply_layers}
\alias{add_blank_rental_supply_layers}
\title{Add blank rental supply layers}
\usage{
add_blank_rental_supply_layers(map)
}
\arguments{
\item{map}{Map created via \link{map_toronto}}
}
\description{
Add a blank layers of \link{rental_supply_by_neighbourhood} to a map (created via \link{map_toronto}). There is one layer for apartments (rental_supply_apartment), condo (rental_supply_condo), and secondary (rental_supply_non_condo). The purpose of this function is to allow for toggling the layers on and off, via \link{toggle_layer_visible} and \link{toggle_layer_invisible}.
}
\examples{
library(sf)
map_toronto() \%>\%
  add_blank_rental_supply_layers() \%>\%
  toggle_layer_visible("rental_supply_condo")
}
