% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{address_points}
\alias{address_points}
\title{Address Points}
\format{
An object of class \code{tbl_Pool} (inherits from \code{tbl_sql}, \code{tbl_lazy}, \code{tbl}) with NA rows and 4 columns.
}
\usage{
address_points
}
\description{
Address points for over 500,000 addresses within the City of Toronto, retrieved from \href{https://open.toronto.ca/dataset/address-points-municipal-toronto-one-address-repository/}{Address Points (Municipal) - Toronto One Address Repository}. The data is stored in a SQLite database, so it can be filtered like a regular data frame, then results must be "collected" with \link[dplyr]{collect}.
}
\examples{
\dontrun{
library(dplyr)
address_points \%>\%
  collect()

address_points \%>\%
  filter(address == "404 Lake Promenade") \%>\%
  collect()
}
}
\keyword{datasets}
