% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{address_points_just_address}
\alias{address_points_just_address}
\title{Address Points addresses only}
\format{
An object of class \code{character} of length 523134.
}
\usage{
address_points_just_address
}
\description{
Addresses only from \link{address_points}
}
\keyword{datasets}
