% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{agi_applications_and_tdf}
\alias{agi_applications_and_tdf}
\title{Above Guideline Increase applications and Tenant Defense Fund grants}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2219 rows and 10 columns.
}
\usage{
agi_applications_and_tdf
}
\description{
The locations of \href{https://www.torontotenants.org/above_guideline_rent_increases}{Above Guideline Increase applications} along with their dates, and whether tenants at those locations received a \href{https://www.toronto.ca/community-people/employment-social-support/housing-support/financial-support-for-renters/}{Tenant Defense Fund} grant to fight the increase. If so, the record also includes how much they were able to reduce the increase by.
}
\keyword{datasets}
