% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amenity_density}
\alias{amenity_density}
\title{Amenity Density}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 12622 rows and 4 columns.
}
\usage{
amenity_density
}
\description{
Amenity Density is a subset of \link{proximity_measures}, showing whether each dissemination block in Toronto is high, medium, or low amenity dense.
}
\keyword{datasets}
