% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amenity_density_by_neighbourhood}
\alias{amenity_density_by_neighbourhood}
\title{Amenity density by neighbourhood}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 420 rows and 3 columns.
}
\usage{
amenity_density_by_neighbourhood
}
\description{
An aggregation of amenity density from \link{proximity_measures}, showing what proportion of a neighbourhood's population lives in high, medium, or low amenity dense areas.
}
\keyword{datasets}
