% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{apartment_building_registry}
\alias{apartment_building_registry}
\title{Apartment building registry}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3479 rows and 75 columns.
}
\usage{
apartment_building_registry
}
\description{
Apartment building registry from City of Toronto, retrieved from \link{https://open.toronto.ca/dataset/apartment-building-registration/} and geocoded using the Bing geocoding API (\link{https://docs.microsoft.com/en-us/bingmaps/rest-services/locations/find-a-location-by-address}). Contains the address of all rental apartment buildings in the city with 3 or more storeys and 10 or more units.
}
\keyword{datasets}
