% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{city_profile}
\alias{city_profile}
\title{City of Toronto Profile}
\format{
An object of class \code{list} of length 11.
}
\usage{
city_profile
}
\description{
Variables describing profiles the City of Toronto Toronto from the 2016 census. Includes average values of variables for the city as well as distributions across neighbourhoods.
}
\details{
The variables are:
\itemize{
\item \strong{Population change} (value and distribution): Population change reflects the per cent change in the population since the last Census in 2011.
\item \strong{Population density} (distribution): Number of people per square kilometer.
\item \strong{Household size} (breakdown): Breakdown of number of persons in private households.
\item \strong{Average total income} (breakdown): Average total income broken down by one and two-plus person households. Total income refers to the total amount of income for a household in 2015 that is of a regular and recurring nature, such as investment and pension income, employment income, and income from government sources.
\item \strong{Unaffordable housing} (value and distribution): Unaffordable housing is the percentage of private households spending more than 30 per cent of their total household income on shelter costs.
\item \strong{Low income (LIM-AT)} (value and distribution): Low income (LIM-AT) is the percentage of people in private households in low income status according to the Low Income Measure, After-Tax.
\item \strong{Visible minority} (breakdown): Visible minority population is the percentage of people in private households who belong to a visible minority group, i.e. persons, other than Aboriginal peoples, who are non-Caucasian in race or non-white in colour.
\item \strong{Private dwellings by structure type} (breakdown): Breakdown of dwellings by structure type, referring to the structural characteristics and/or configuration of the dwelling. Two categories in the Census, mobile dwellings and ""other"" single-attached house, make up a very small proportion of Toronto's occupied dwellings and so are not reported.
\item \strong{Number of bedrooms} (breakdown): Breakdown of number of bedrooms.
\item \strong{Household tenure} (breakdown): Households by tenure provides data on the number and proportion of private households that own or rent their dwelling. A third tenure type, which is not present in the City of Toronto, is whether the dwelling is band housing (on an Indian reserve or settlement).
\item \strong{Average shelter cost for rentals} (value and distribution): Shelter cost refers to the average monthly total of all shelter expenses paid by households that rent their dwelling. Shelter costs include, where applicable, the rent and the costs of electricity, heat, water and other municipal services. The reference period for shelter cost data is 2016, while household total income is reported for the year 2015.
}
}
\examples{
city_profile[["average_total_income"]]
}
\keyword{datasets}
