% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_and_definitions}
\alias{data_and_definitions}
\title{Data and definitions}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 18 rows and 6 columns.
}
\usage{
data_and_definitions
}
\description{
Definitions and sources of data used throughout the app.
}
\keyword{datasets}
