% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_neighbourhood_profile.R
\name{display_neighbourhood_profile}
\alias{display_neighbourhood_profile}
\title{Plot a neighbourhood profile variable}
\usage{
display_neighbourhood_profile(
  data,
  variable,
  compare = TRUE,
  width = 20,
  dollar = FALSE,
  type = "plot",
  static = FALSE
)
}
\arguments{
\item{data}{Neighbourhood profiles data for a given neighbourhood, from \link{neighbourhood_aggregate}, or for the city, from \link{city_aggregate}}

\item{variable}{Variable to visualize.}

\item{compare}{Whether to compare to City of Toronto values. Defaults to TRUE. FALSE is useful when you want to plot *just* the values for as neighbourhood or *just* the values for the city.}

\item{width}{Passed along to str_wrap for wrapping y-axis labels. Defaults to a width of 20.}

\item{dollar}{Whether the variable shown is in dollars. Defaults to FALSE.}

\item{type}{Type of display, "plot" or "table". Defaults to "plot".}
}
\description{
Plot a neighbourhood profile variable
}
\examples{
{
  neighbourhood_aggregate[["Danforth"]] \%>\%
    display_neighbourhood_profile("household_size")

  neighbourhood_aggregate[["Danforth"]] \%>\%
    display_neighbourhood_profile("average_total_income")
}
}
