% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eviction_hearings}
\alias{eviction_hearings}
\title{Evictions Hearings}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1916 rows and 6 columns.
}
\usage{
eviction_hearings
}
\description{
Locations (and number) of eviction hearings scheduled by the Landlord Tenant Board between November 2, 2020 to January 31, 2021.
}
\keyword{datasets}
