% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_address.R
\name{geocode_address}
\alias{geocode_address}
\title{Geocode an address}
\usage{
geocode_address(
  address,
  base = "http://dev.virtualearth.net/REST/v1/Locations/CA/",
  token = Sys.getenv("BING_TOKEN"),
  quiet = FALSE
)
}
\arguments{
\item{address}{Address to geocode}

\item{base}{Base URL of Bing Geocoder: http://dev.virtualearth.net/REST/v1/Locations/CA/}

\item{token}{BING Geocoder token, defaults to \code{BING_TOKEN} environment variable.}

\item{quiet}{Whether to silence message that address is being geocoded, with its status - defaults to \code{FALSE} (display the message). Useful if geocoding many addresses at once.}
}
\value{
A tibble containing status code, address, municipality, postal code, method used, method confidence, latitude, and longitude.
}
\description{
Geocode an address using the \href{https://docs.microsoft.com/en-us/bingmaps/rest-services/locations/find-a-location-by-address}{Bing geocoding API}, to get a cleaned address, municipality, postal code, latitude, and longitude. This function adds an additional 0.25 seconds at the end of the call to comply with the \href{http://mapsforenterprise.binginternal.com/en-us/maps/product}{Bing API license}, which only allows 5 calls per second.
}
\examples{
geocode_address("235 Bloor St E")
}
