% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neighbourhood_profiles}
\alias{neighbourhood_profiles}
\title{City of Toronto Neighbourhood Profiles}
\format{
An object of class \code{list} of length 140.
}
\usage{
neighbourhood_profiles
}
\description{
Variables describing profiles of neighbourhoods in Toronto. The data set is a list, with one element for each neighbourhood, then within each neighbourhood one element for each variable.
}
\examples{
neighbourhood_profiles[["Danforth"]]
}
\keyword{datasets}
