% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neighbourhoods}
\alias{neighbourhoods}
\title{City of Toronto Neighbourhoods
Neighbourhood boundaries from the City of Toronto, from \href{https://open.toronto.ca/dataset/neighbourhoods/}{Neighbourhoods data set}}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 140 rows and 3 columns.
}
\usage{
neighbourhoods
}
\description{
City of Toronto Neighbourhoods
Neighbourhood boundaries from the City of Toronto, from \href{https://open.toronto.ca/dataset/neighbourhoods/}{Neighbourhoods data set}
}
\examples{
library(sf)
neighbourhoods
}
\keyword{datasets}
