% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_age_pyramid.R
\name{plot_age_pyramid}
\alias{plot_age_pyramid}
\title{Plot age pyramid}
\usage{
plot_age_pyramid(data, horizontal = FALSE)
}
\arguments{
\item{data}{Input age pyramid data, with columns \code{sex}, \code{metric}, \code{age_group}, \code{value}.}

\item{horizontal}{Whether the plot should be horizontal (with age groups on the x-axis and proportions on the y-axis). Defaults to \code{FALSE} (age groups on the y-axis and proportions on the x-axis).}
}
\description{
Plot age pyramid
}
\examples{
{
neighbourhood_profiles[["Danforth"]][["age_pyramid"]] \%>\%
  plot_age_pyramid()
}
}
