% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_household_size.R
\name{plot_household_size}
\alias{plot_household_size}
\title{Plot household size}
\usage{
plot_household_size(data)
}
\arguments{
\item{data}{Input household size data for a neighbourhood, from \link{neighbourhood_profiles}.}
}
\description{
Plot household size
}
\examples{
{
neighbourhood_profiles[["Danforth"]][["household_size"]] \%>\%
  plot_household_size()
}
}
