% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_lemur.R
\name{theme_lemur}
\alias{theme_lemur}
\title{LEMUR ggplot2 theme}
\usage{
theme_lemur(base_size = 14, base_family = "Lato")
}
\arguments{
\item{base_size}{Base font size}

\item{base_family}{Base font family}
}
\description{
LEMUR ggplot2 theme
}
