% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggle_layer.R
\name{toggle_layer_visible}
\alias{toggle_layer_visible}
\alias{toggle_layer_invisible}
\title{Toggle mapbox layer visible or invisible}
\usage{
toggle_layer_visible(map, id)

toggle_layer_invisible(map, id)
}
\arguments{
\item{map}{Map e.g. from \link{map_toronto}}

\item{id}{ID of existing layer, added via e.g. \link{add_blank_apartment_layer}}
}
\description{
Toggle mapbox layer visible or invisible
}
\examples{
library(sf)

map_toronto() \%>\%
  add_blank_amenity_density_layer() \%>\%
  toggle_layer_visible("amenity_density")
}
