% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{total_affordable_by_neighbourhood}
\alias{total_affordable_by_neighbourhood}
\title{Total Affordable By Neighbourhood}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 140 rows and 4 columns.
}
\usage{
total_affordable_by_neighbourhood
}
\description{
An estimate of the available number of affordable (deeply and very) rental units by neighbourhood.
}
\keyword{datasets}
