% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom_map_to_address.R
\name{zoom_map_to_address}
\alias{zoom_map_to_address}
\title{Zoom map of Toronto to address}
\usage{
zoom_map_to_address(map, address)
}
\arguments{
\item{map}{Map created by \link{map_toronto}}

\item{address}{Address or sf tibble with geometry column}
}
\description{
Zoom map of Toronto to address
}
\examples{
# library(sf)

# map_toronto() \%>\%
#  add_blank_address_layer() \%>\%
#  zoom_map_to_address("378 Markham St")
}
