% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neighbourhoods}
\alias{neighbourhoods}
\title{City of Toronto Neighbourhoods
Neighbourhood boundaries from the City of Toronto, retrieved from \link{https://open.toronto.ca/dataset/neighbourhoods/}}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 140 rows and 2 columns.
}
\usage{
neighbourhoods
}
\description{
City of Toronto Neighbourhoods
Neighbourhood boundaries from the City of Toronto, retrieved from \link{https://open.toronto.ca/dataset/neighbourhoods/}
}
\keyword{datasets}
