% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_neighbourhood_profile.R
\name{plot_neighbourhood_profile}
\alias{plot_neighbourhood_profile}
\title{Plot a neighbourhood profile variable}
\usage{
plot_neighbourhood_profile(data, variable)
}
\arguments{
\item{data}{Neighbourhood profiles data for a given neighbourhood, from \link{neighbourhood_profiles}.}

\item{variablle}{Variable to visualize.}
}
\description{
Plot a neighbourhood profile variable
}
\examples{
{
neighbourhood_profiles[["Danforth"]] \%>\%
  plot_neighbourhood_profile("household_size")

neighbourhood_profiles[["Danforth"]] \%>\%
  plot_neighbourhood_profile("average_total_income")
}
}
