% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_neighbourhood_profile.R
\name{plot_neighbourhood_profile_distribution}
\alias{plot_neighbourhood_profile_distribution}
\title{Plot the distribution of a neighbourhood profile variable}
\usage{
plot_neighbourhood_profile_distribution(data, variable)
}
\arguments{
\item{data}{Neighbourhood profiles data for a given neighbourhood, from \link{neighbourhood_profiles}.}

\item{variable}{Variable to visualize}
}
\description{
Plot the distribution of a variable, across neighbourhoods, with a line showing the current neighbourhood's value
}
\examples{
neighbourhood_profiles[["Danforth"]] \%>\%
  plot_neighbourhood_profile_distribution("population_density")
}
