% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toronto}
\alias{toronto}
\title{City of Toronto boundaries}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 4 columns.
}
\usage{
toronto
}
\description{
City of Toronto boundaries, retrieved from \link{https://open.toronto.ca/dataset/regional-municipal-boundary/}
}
\keyword{datasets}
